refactor: address CodeRabbit review suggestions
